backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
python-multipart>=0.0.5
openai>=0.27.0
SpeechRecognition>=3.8.1
PyAudio>=0.2.11
python-dotenv>=0.19.0